@@ -3,22 +3,21 @@ package com.example.sharingapp;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Bitmap;
+import android.os.Handler;
 import android.provider.MediaStore;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.ArrayAdapter;
+import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageButton;
 import android.widget.ImageView;
-import android.widget.Spinner;
-import android.widget.Switch;
 import android.widget.TextView;
+import android.widget.Toast;
 
 /**
  * Editing a pre-existing item consists of deleting the old item and adding a new item with the old
  * item's id.
- * Note: invisible EditText is used to setError for status. For whatever reason we cannot .setError to
- * the status Switch so instead an error is set to an "invisible" EditText.
  */
 public class EditItemActivity extends AppCompatActivity implements Observer {
 
@@ -30,8 +29,8 @@ public class EditItemActivity extends AppCompatActivity implements Observer {
 
     private Context context;
 
-    private ContactList contact_list = new ContactList();
-    private ContactListController contact_list_controller = new ContactListController(contact_list);
+    private UserList user_list = new UserList();
+    private UserListController user_list_controller = new UserListController(user_list);
 
     private Bitmap image;
     private int REQUEST_CODE = 1;
@@ -43,48 +42,74 @@ public class EditItemActivity extends AppCompatActivity implements Observer {
     private EditText length;
     private EditText width;
     private EditText height;
-    private Spinner borrower_spinner;
-    private TextView  borrower_tv;
-    private Switch status;
-    private EditText invisible;
+    private EditText minimum_bid;
 
-    private ArrayAdapter<String> adapter;
-    private boolean on_create_update = false;
+    private TextView borrower_left_tv;
+    private TextView borrower_right_tv;
+    private TextView status_right_tv;
+
+    private Button save_button;
+    private Button view_bids_button;
+    private Button contact_info_button;
+    private Button delete_button;
+    private Button set_available_button;
+    private ImageButton add_image_button;
+    private ImageButton delete_image_button;
+
+    private boolean on_create_update;
     private int pos;
+
+    private String title_str;
+    private String maker_str;
+    private String description_str;
+    private String length_str;
+    private String width_str;
+    private String height_str;
+    private String user_id;
+    private String minimum_bid_str;
+    private String status_str;
+    private String borrower_username_str;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_edit_item);
 
-        title = findViewById(R.id.title);
-        maker = findViewById(R.id.maker);
-        description = findViewById(R.id.description);
-        length = findViewById(R.id.length);
-        width = findViewById(R.id.width);
-        height = findViewById(R.id.height);
-        borrower_spinner = findViewById(R.id.borrower_spinner);
-        borrower_tv = findViewById(R.id.borrower_tv);
-        photo = findViewById(R.id.image_view);
-        status = findViewById(R.id.available_switch);
-        invisible = findViewById(R.id.invisible);
+        title = (EditText) findViewById(R.id.title);
+        maker = (EditText) findViewById(R.id.maker);
+        description = (EditText) findViewById(R.id.description);
+        length = (EditText) findViewById(R.id.length);
+        width = (EditText) findViewById(R.id.width);
+        height = (EditText) findViewById(R.id.height);
+        status_right_tv = (TextView) findViewById(R.id.status_right_tv);
+        minimum_bid = (EditText) findViewById(R.id.minimum_bid);                    // initially GONE
+        borrower_left_tv = (TextView) findViewById(R.id.borrower_left_tv);          // initially GONE
+        borrower_right_tv = (TextView) findViewById(R.id.borrower_right_tv);        // initially GONE
+        photo = (ImageView) findViewById(R.id.image_view);
 
-        invisible.setVisibility(View.GONE);
+        add_image_button = (ImageButton) findViewById(R.id.add_image_button);       // initially GONE
+        delete_image_button = (ImageButton) findViewById(R.id.cancel_image_button); // initially GONE
+        delete_button = (Button) findViewById(R.id.delete_item);                    // initially GONE
+        save_button = (Button) findViewById(R.id.save_button);                      // initially GONE
+        view_bids_button = (Button) findViewById(R.id.view_bids_button);            // initially GONE
+        set_available_button = (Button) findViewById(R.id.set_available_button);    // initially GONE
+        contact_info_button = (Button) findViewById(R.id.contact_info_button);      // initially GONE
 
         Intent intent = getIntent(); // Get intent from ItemsFragment
         pos = intent.getIntExtra("position", 0);
+        user_id = intent.getStringExtra("user_id");
 
         context = getApplicationContext();
 
+        on_create_update = false; // Suppress first call to update()
         item_list_controller.addObserver(this);
         item_list_controller.loadItems(context);
 
         on_create_update = true;
+        user_list_controller.addObserver(this);
+        user_list_controller.loadUsers(context); // Call to update occurs
 
-        contact_list_controller.addObserver(this);
-        contact_list_controller.loadContacts(context);
-
-        on_create_update = false;
+        on_create_update = false; // Suppress any further calls to update()
     }
 
     public void addPhoto(View view) {
@@ -97,6 +122,7 @@ public class EditItemActivity extends AppCompatActivity implements Observer {
     public void deletePhoto(View view) {
         image = null;
         photo.setImageResource(android.R.drawable.ic_menu_gallery);
+        Toast.makeText(context, "Photo removed.", Toast.LENGTH_SHORT).show();
     }
 
     @Override
@@ -105,124 +131,113 @@ public class EditItemActivity extends AppCompatActivity implements Observer {
             Bundle extras = intent.getExtras();
             image = (Bitmap) extras.get("data");
             photo.setImageBitmap(image);
+            Toast.makeText(context, "Photo added.", Toast.LENGTH_SHORT).show();
         }
+    }
+
+    @Override
+    public void onBackPressed() {
+        Intent main_intent = new Intent(this, MainActivity.class);
+        main_intent.putExtra("user_id", user_id);
+        startActivity(main_intent);
     }
 
     public void deleteItem(View view) {
-
-        // Delete item
         boolean success = item_list_controller.deleteItem(item, context);
-        if (!success) {
+        if (!success){
             return;
         }
 
         // End EditItemActivity
         item_list_controller.removeObserver(this);
+        user_list_controller.removeObserver(this);
 
-        Intent intent = new Intent(this, MainActivity.class);
-        startActivity(intent);
+        final Intent intent = new Intent(this, MainActivity.class);
+        intent.putExtra("user_id", user_id);
+
+        // Delay launch of new activity to allow server more time to process request
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Toast.makeText(context, "Item removed.", Toast.LENGTH_SHORT).show();
+                startActivity(intent);
+            }
+        }, 750);
     }
 
     public void saveItem(View view) {
+        title_str = title.getText().toString();
+        maker_str = maker.getText().toString();
+        description_str = description.getText().toString();
+        length_str = length.getText().toString();
+        width_str = width.getText().toString();
+        height_str = height.getText().toString();
+        status_str = item_controller.getStatus();
+        minimum_bid_str = minimum_bid.getText().toString();
 
-        if (!validateInput()) {
+        if(!validateInput()){
             return;
         }
 
-        Contact contact = null;
+        String item_id = item_controller.getId(); // Reuse the item id
 
-        if (!status.isChecked()) {
-            String borrower_str = borrower_spinner.getSelectedItem().toString();
-            contact = contact_list_controller.getContactByUsername(borrower_str);
-        }
-
-        String id = item_controller.getId(); // Reuse the item id
-        Item updated_item = new Item(title_str, maker_str, description_str, image, id);
+        Item updated_item = new Item(title_str, maker_str, description_str, user_id, minimum_bid_str, image, item_id);
         ItemController updated_item_controller = new ItemController(updated_item);
         updated_item_controller.setDimensions(length_str, width_str, height_str);
+        updated_item_controller.setStatus(status_str);
 
-        boolean checked = status.isChecked();
-        if (!checked) {
-            updated_item_controller.setStatus("Borrowed");
-            updated_item_controller.setBorrower(contact);
-        }
-
-        // Edit item
         boolean success = item_list_controller.editItem(item, updated_item, context);
-        if (!success) {
+        if (!success){
             return;
         }
 
-        // End EditItemActivity
         item_list_controller.removeObserver(this);
+        user_list_controller.removeObserver(this);
 
-        Intent intent = new Intent(this, MainActivity.class);
+        // End EditItemActivity
+        final Intent intent = new Intent(this, MainActivity.class);
+        intent.putExtra("user_id", user_id);
+        Toast.makeText(context, "Item saved.", Toast.LENGTH_SHORT).show();
         startActivity(intent);
     }
 
-    /**
-     * Checked == "Available"
-     * Unchecked == "Borrowed"
-     */
-    public void toggleSwitch(View view){
-        if (status.isChecked()) {
-            // Means was previously borrowed, switch was toggled to available
-            borrower_spinner.setVisibility(View.GONE);
-            borrower_tv.setVisibility(View.GONE);
-            item_controller.setBorrower(null);
-            item_controller.setStatus("Available");
+    public void viewBids(View view){
+        user_list_controller.removeObserver(this);
+        item_list_controller.removeObserver(this);
 
-        } else {
-            // Means not borrowed
-            if (contact_list.getSize()==0){
-                // No contacts, need to add contacts to be able to add a borrower
-                invisible.setEnabled(false);
-                invisible.setVisibility(View.VISIBLE);
-                invisible.requestFocus();
-                invisible.setError("No contacts available! Must add borrower to contacts.");
-                status.setChecked(true); // Set switch to available
+        Intent intent = new Intent(this, ViewItemBidsActivity.class);
+        intent.putExtra("user_id", user_id);
+        intent.putExtra("item_id", item_controller.getId());
+        startActivity(intent);
+    }
 
-            } else {
-                // Means was previously available
-                borrower_spinner.setVisibility(View.VISIBLE);
-                borrower_tv.setVisibility(View.VISIBLE);
-            }
-        }
+    public void setAvailable(View view){
+        item_controller.setStatus("Available"); // Update status
+        saveItem(view); // Must save the item so that the change in status is saved
     }
 
     /**
-     * Only need to update the view from the onCreate method
+     * Only need to update the view once from the onCreate method
      */
     public void update() {
         if (on_create_update){
-            adapter = new ArrayAdapter<String>(this, android.R.layout.simple_spinner_dropdown_item,
-                    contact_list_controller.getAllUsernames());
-            borrower_spinner.setAdapter(adapter);
 
+            // For all status options we do the following
             item = item_list_controller.getItem(pos);
             item_controller = new ItemController(item);
-
-            Contact contact = item_controller.getBorrower();
-            if (contact != null){
-                int contact_pos = contact_list_controller.getIndex(contact);
-                borrower_spinner.setSelection(contact_pos);
-            }
 
             title.setText(item_controller.getTitle());
             maker.setText(item_controller.getMaker());
             description.setText(item_controller.getDescription());
-
             length.setText(item_controller.getLength());
             width.setText(item_controller.getWidth());
             height.setText(item_controller.getHeight());
 
-            String status_str = item_controller.getStatus();
-            if (status_str.equals("Borrowed")) {
-                status.setChecked(false);
-            } else {
-                borrower_tv.setVisibility(View.GONE);
-                borrower_spinner.setVisibility(View.GONE);
-            }
+            status_str = item_controller.getStatus();
+            status_right_tv.setText(status_str);
+
+            minimum_bid_str = item_controller.getMinBid().toString();
+            minimum_bid.setText(minimum_bid_str);
 
             image = item_controller.getImage();
             if (image != null) {
@@ -230,17 +245,61 @@ public class EditItemActivity extends AppCompatActivity implements Observer {
             } else {
                 photo.setImageResource(android.R.drawable.ic_menu_gallery);
             }
+
+            // AVAILABLE
+            if (status_str.equals("Available")){
+
+                title.setEnabled(true);
+                maker.setEnabled(true);
+                description.setEnabled(true);
+                length.setEnabled(true);
+                width.setEnabled(true);
+                height.setEnabled(true);
+                minimum_bid.setEnabled(true);
+
+                add_image_button.setVisibility(View.VISIBLE);
+                delete_image_button.setVisibility(View.VISIBLE);
+                delete_button.setVisibility(View.VISIBLE);
+                save_button.setVisibility(View.VISIBLE);
+            }
+
+            // BIDDED
+            if (status_str.equals("Bidded")){
+
+                view_bids_button.setVisibility(View.VISIBLE);
+                title.setEnabled(false);
+                maker.setEnabled(false);
+                description.setEnabled(false);
+                length.setEnabled(false);
+                width.setEnabled(false);
+                height.setEnabled(false);
+                minimum_bid.setEnabled(false);
+            }
+
+            // BORROWED
+            if (status_str.equals("Borrowed")){
+
+                title.setEnabled(false);
+                maker.setEnabled(false);
+                description.setEnabled(false);
+                length.setEnabled(false);
+                width.setEnabled(false);
+                height.setEnabled(false);
+                minimum_bid.setEnabled(false);
+
+                borrower_left_tv.setVisibility(View.VISIBLE);
+                borrower_right_tv.setVisibility(View.VISIBLE);
+                contact_info_button.setVisibility(View.VISIBLE);
+                set_available_button.setVisibility(View.VISIBLE);
+
+                User borrower = item_controller.getBorrower();
+                borrower_username_str = borrower.getUsername();
+                borrower_right_tv.setText(borrower_username_str);
+            }
         }
     }
 
-    protected boolean validateInput() {
-        String title_str = title.getText().toString();
-        String maker_str = maker.getText().toString();
-        String description_str = description.getText().toString();
-        String length_str = length.getText().toString();
-        String width_str = width.getText().toString();
-        String height_str = height.getText().toString();
-
+    public boolean validateInput(){
         if (title_str.equals("")) {
             title.setError("Empty field!");
             return false;
@@ -271,6 +330,25 @@ public class EditItemActivity extends AppCompatActivity implements Observer {
             return false;
         }
 
+        if (minimum_bid_str.equals("")) {
+            minimum_bid.setError("Empty field!");
+            return false;
+        }
+
+        if (Float.valueOf(minimum_bid_str) <= 0) {
+            minimum_bid.setError("Starting bid must be above 0!");
+            return false;
+        }
+
         return true;
+    }
+
+    public void viewUserActivity(View view){
+        user_list_controller.removeObserver(this);
+        item_list_controller.removeObserver(this);
+
+        Intent intent = new Intent(this, ViewUserActivity.class);
+        intent.putExtra("borrower_username_str", borrower_username_str);
+        startActivity(intent);
     }
 }
