@@ -3,7 +3,7 @@ package com.example.sharingapp;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,14 +14,16 @@ import android.widget.ListView;
 import java.util.ArrayList;
 
 /**
- * Superclass of AvailableItemsFragment, BorrowedItemsFragment and AllItemsFragment
+ * Superclass of AvailableItemsFragment, BorrowedItemsFragment, BiddedItemsFragment and AllItemsFragment
  */
 public abstract class ItemsFragment extends Fragment implements Observer {
 
     private ItemList item_list = new ItemList();
-    ItemListController item_list_controller = new ItemListController(item_list);
 
+    ItemListController item_list_controller = new ItemListController(item_list);
     View rootView;
+    String user_id;
+
     private ListView list_view;
     private ArrayAdapter<Item> adapter;
     private ArrayList<Item> selected_items;
@@ -33,11 +35,10 @@ public abstract class ItemsFragment extends Fragment implements Observer {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        context = getContext();
+        this.context = getContext();
 
-        // Don't update view yet. Wait until after items have been filtered.
-        item_list_controller.loadItems(context);
-        update = true;
+        item_list_controller.loadItems(context); // Call to update() suppressed
+        update = true; // Future calls to update() permitted
 
         this.inflater = inflater;
         this.container = container;
@@ -47,8 +48,12 @@ public abstract class ItemsFragment extends Fragment implements Observer {
 
     public void setVariables(int resource, int id ) {
         rootView = inflater.inflate(resource, container, false);
-        list_view = rootView.findViewById(id);
+        list_view = (ListView) rootView.findViewById(id);
         selected_items = filterItems();
+    }
+
+    public void setUserId(Bundle b) {
+        this.user_id = b.getString("user_id", user_id);
     }
 
     public void loadItems(Fragment fragment){
@@ -58,19 +63,18 @@ public abstract class ItemsFragment extends Fragment implements Observer {
     }
 
     public void setFragmentOnItemLongClickListener(){
-
         // When item is long clicked, this starts EditItemActivity
         list_view.setOnItemLongClickListener(new android.widget.AdapterView.OnItemLongClickListener() {
 
             @Override
             public boolean onItemLongClick(AdapterView<?> parent, View view, int pos, long id) {
-
                 Item item = adapter.getItem(pos);
 
                 int meta_pos = item_list_controller.getIndex(item);
                 if (meta_pos >= 0) {
 
                     Intent edit = new Intent(context, EditItemActivity.class);
+                    edit.putExtra("user_id", user_id);
                     edit.putExtra("position", meta_pos);
                     startActivity(edit);
                 }
@@ -80,7 +84,8 @@ public abstract class ItemsFragment extends Fragment implements Observer {
     }
 
     /**
-     * filterItems is implemented independently by AvailableItemsFragment, BorrowedItemsFragment and AllItemsFragment
+     * filterItems is implemented independently by AvailableItemsFragment, BorrowedItemsFragment
+     * BiddedItemsFragment and AllItemsFragment
      * @return selected_items
      */
     public abstract ArrayList<Item> filterItems();
@@ -99,7 +104,8 @@ public abstract class ItemsFragment extends Fragment implements Observer {
      */
     public void update(){
         if (update) {
-            adapter = new ItemAdapter(context, selected_items, fragment);
+            selected_items = filterItems(); // Ensure items are filtered
+            adapter = new ItemFragmentAdapter(context, selected_items, fragment);
             list_view.setAdapter(adapter);
             adapter.notifyDataSetChanged();
         }
