@@ -1,8 +1,9 @@
 package com.example.sharingapp;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import android.os.Bundle;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 /**
  * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
@@ -10,21 +11,33 @@ import android.support.v4.app.FragmentPagerAdapter;
  */
 public class SectionsPagerAdapter extends FragmentPagerAdapter {
 
-    public SectionsPagerAdapter(FragmentManager fm) {
+    private String user_id;
+
+    public SectionsPagerAdapter(FragmentManager fm, String user_id) {
         super(fm);
+        this.user_id = user_id;
     }
 
     @Override
     public Fragment getItem(int position) {
+        Bundle bundle = new Bundle();
+        bundle.putString("user_id", user_id);
         switch (position) {
             case 0:
                 AllItemsFragment all_items_fragment = new AllItemsFragment();
+                all_items_fragment.setArguments(bundle);
                 return all_items_fragment;
             case 1:
                 AvailableItemsFragment available_items_fragment = new AvailableItemsFragment();
+                available_items_fragment.setArguments(bundle);
                 return available_items_fragment;
             case 2:
+                BiddedItemsFragment bidded_items_fragment = new BiddedItemsFragment();
+                bidded_items_fragment.setArguments(bundle);
+                return bidded_items_fragment;
+            case 3:
                 BorrowedItemsFragment borrowed_items_fragment = new BorrowedItemsFragment();
+                borrowed_items_fragment.setArguments(bundle);
                 return borrowed_items_fragment;
             default:
                 return null;
@@ -33,7 +46,7 @@ public class SectionsPagerAdapter extends FragmentPagerAdapter {
 
     @Override
     public int getCount() {
-        return 3; // Three pages
+        return 4;  // Four pages
     }
 
     @Override
@@ -44,7 +57,10 @@ public class SectionsPagerAdapter extends FragmentPagerAdapter {
             case 1:
                 return "Available";
             case 2:
+                return "Bidded";
+            case 3:
                 return "Borrowed";
+
         }
         return null;
     }
