@@ -8,14 +8,14 @@ import android.view.ViewGroup;
 import java.util.ArrayList;
 
 /**
- * Displays a list of all "Borrowed" Items
+ * Displays a list of all "Borrowed" items
  */
 public class BorrowedItemsFragment extends ItemsFragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-
         super.onCreateView(inflater,container, savedInstanceState);
+        super.setUserId(this.getArguments());
         super.setVariables(R.layout.borrowed_items_fragment, R.id.my_borrowed_items);
         super.loadItems(BorrowedItemsFragment.this);
         super.setFragmentOnItemLongClickListener();
@@ -25,6 +25,6 @@ public class BorrowedItemsFragment extends ItemsFragment {
 
     public ArrayList<Item> filterItems() {
         String status = "Borrowed";
-        return item_list_controller.filterItemsByStatus(status);
+        return item_list_controller.filterItems(user_id, status);
     }
 }
