@@ -17,17 +17,21 @@ public class Item extends Observable {
     private String description;
     private Dimensions dimensions;
     private String status;
-    private Contact borrower;
+    private Float minimum_bid;
+    private User borrower;
+    private String owner_id;
     protected transient Bitmap image;
     protected String image_base64;
     private String id;
 
-    public Item(String title, String maker, String description, Bitmap image, String id) {
+    public Item(String title, String maker, String description, String owner_id, String minimum_bid, Bitmap image, String id) {
         this.title = title;
         this.maker = maker;
         this.description = description;
         this.dimensions = null;
+        this.owner_id = owner_id;
         this.status = "Available";
+        this.minimum_bid = Float.valueOf(minimum_bid);
         this.borrower = null;
         addImage(image);
 
@@ -79,8 +83,26 @@ public class Item extends Observable {
         return description;
     }
 
+    public Float getMinBid() {
+        return this.minimum_bid;
+    }
+
+    public void setMinBid(Float minimum_bid) {
+        this.minimum_bid = minimum_bid;
+        notifyObservers();
+    }
+
+    public void setOwnerId(String owner_id) {
+        this.owner_id = owner_id;
+        notifyObservers();
+    }
+
+    public String getOwnerId() {
+        return owner_id;
+    }
+
     public void setDimensions(String length, String width, String height) {
-        this.dimensions = new Dimensions(length, width, height);
+        dimensions = new Dimensions(length, width, height);
         notifyObservers();
     }
 
@@ -105,13 +127,20 @@ public class Item extends Observable {
         return status;
     }
 
-    public void setBorrower(Contact borrower) {
+    public void setBorrower(User borrower) {
         this.borrower = borrower;
         notifyObservers();
     }
 
-    public Contact getBorrower() {
+    public User getBorrower() {
         return borrower;
+    }
+
+    public String getBorrowerUsername() {
+        if (borrower != null){
+            return borrower.getUsername();
+        }
+        return null;
     }
 
     public void addImage(Bitmap new_image){
