@@ -4,5 +4,5 @@ package com.example.sharingapp;
  * Observer Interface
  */
 public interface Observer {
-    void update();
+    public void update();
 }
